requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pymongo[srv]==4.5.0
pyTelegramBotAPI==4.10.0
pytz==2023.3
requests==2.31.0
certifi==2023.7.22
